fix: attach delete lib all fix

+++ v13/libs/attachDelete.ts
@@ -5,13 +5,17 @@ export async function attachDeletable(m: Message, id: string) {
   await m.react(emoji).catch(() => {});
   m.awaitReactions({
     time: 120000,
+    max: 1,
     dispose: true,
     filter(r, u) {
-      return r.emoji.identifier == emoji && u.id == id;
+      return r.emoji.toString() == emoji && u.id == id;
     },
-  }).then(() => {
-    m.delete().catch(() => {});
-    m.reactions.removeAll().catch(() => {});
-  });
+  })
+    .then((c) => {
+      m.reactions.removeAll().catch(() => {});
+      if (c.size == 0) return;
+      m.delete().catch(() => {});
+    })
+    .catch(() => {});
   return m;
 }
